pyproject.toml
[project]
name = "companion-server"
version = "0.1.0"
description = "AI Memory-driven Chat Agent Backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "prisma>=0.15.0",
    "redis[hiredis]>=5.2.0",
    "neo4j>=5.27.0",
    "langchain-core>=0.3.0",
    "langchain-ollama>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "sse-starlette>=2.2.0",
    "apscheduler>=3.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
